First commit
The code does the following procedure:
1.The image is first converted to a grayscale image and resized to 256x256 for standard procedures.
2.Since DCT is designed to work for the range (-127,128), the image is levelled by subtracting 128 from each pixel value.
3.The image is now broken down into 8x8 matrices and the DCT for 8x8 block is applied to each.
4.Each block is subjected to compression based on the quantization matrix used.
5.The compressed matrix is converted into a vector and encoded in the form:
[element, no. of zeros occurring before that element]
6.This encoded data is passed through Huffman coding and then compressed further by gzip.
The file is then transmitted on the channel.
7.On the receiver end, all the transforms are applied in reverse to get back the original image.

The main.py will quantize and decquantize the image (without the use of python), hence the code needs to be seggragated accordingly before use

The vector containing the quantized values is to be exported as a .mat file to to the matlab code-zipper.py or huffman.py both of which shall compress the descriptor information.